+++ labview-interop/src/types/array/mod.rs
@@ -4,16 +4,34 @@
 
 #[cfg(feature = "ndarray")]
 mod ndarray;
+#[cfg(feature = "link")]
+mod memory;
 
 use crate::labview_layout;
 use crate::memory::UHandle;
+
+#[repr(transparent)]
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub struct LvArrayDims<const D: usize>([i32; D]);
+
+impl<const D: usize> LvArrayDims<D> {
+    pub fn element_count(&self) -> usize {
+        self.0.iter().fold(1, |size, dim| size * *dim as usize)
+    }
+}
+
+impl<const D: usize> From<[i32; D]> for LvArrayDims<D> {
+    fn from(dim_sizes: [i32; D]) -> Self {
+        Self(dim_sizes)
+    }
+}
 
 labview_layout!(
     /// Internal LabVIEW array representation.
     ///
     /// todo: does this follow cluster packing rules? yes but lots breaks.
     pub struct LVArray<const D: usize, T> {
-        dim_sizes: [i32; D],
+        dim_sizes: LvArrayDims<D>,
         // For 64 bit use the DST syntax which is more correct to what we
         // are representing.
         #[cfg(target_pointer_width = "64")]
@@ -33,7 +51,7 @@ impl<const D: usize, T> LVArray<D, T> {
         let mut size: usize = 1;
 
         for index in 0..D {
-            let element_ptr = std::ptr::addr_of!(self.dim_sizes[index]);
+            let element_ptr = std::ptr::addr_of!(self.dim_sizes.0[index]);
             // Safety: the indexes must be in range due to the const generic value.
             let dim_size = unsafe { std::ptr::read_unaligned(element_ptr) };
             size *= dim_size as usize;
@@ -63,8 +81,7 @@ impl<const D: usize, T> LVArray<D, T> {
 impl<const D: usize, T> LVArray<D, T> {
     /// Get the total number of elements in the array across all dimensions.
     pub fn element_count(&self) -> usize {
-        let size = self.dim_sizes.iter().fold(1, |size, dim| size * dim);
-        size as usize
+       self.dim_sizes.element_count()
     }
 
     /// Get the data component as a slice.
@@ -94,3 +111,26 @@ impl<const D: usize, T> LVArray<D, T> {
 
 /// Definition of a handle to an array. Helper for FFI definitin.
 pub type LVArrayHandle<const D: usize, T> = UHandle<LVArray<D, T>>;
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    #[test]
+    fn dimension_element_count() {
+        let dims = LvArrayDims::<3>([2, 3, 4]);
+        assert_eq!(dims.element_count(), 24);
+
+        let dims = LvArrayDims::<2>([2, 3]);
+        assert_eq!(dims.element_count(), 6);
+
+        let dims = LvArrayDims::<1>([2]);
+        assert_eq!(dims.element_count(), 2);
+    }
+
+    #[test]
+    fn test_dim_equality() {
+        let dims1 = LvArrayDims::<3>([2, 3, 4]);
+        let dims2 = LvArrayDims::<3>([2, 3, 4]);
+        assert_eq!(dims1, dims2);
+    }
+}