+++ labview-test-library/src/error_clusters.rs
@@ -1,4 +1,4 @@
-use labview_interop::types::LVStatusCode;
+use labview_interop::types::{LStrHandle, LVStatusCode};
 /// A simple type for testing the error integration.
 ///
 use labview_interop::types::{ErrorClusterPtr, ToLvError};
@@ -20,4 +20,11 @@ impl ToLvError for ErrorText {
 pub extern "C" fn set_error_cluster(error_cluster: ErrorClusterPtr) -> LVStatusCode {
     let error = ErrorText("This is a test");
     error.write_error(error_cluster).into()
+}
+
+#[no_mangle]
+pub extern "C" fn get_error_description(code: i32, mut string: LStrHandle) {
+    let status = LVStatusCode::from(code);
+    let description = status.description();
+    string.set_str(description.as_ref()).unwrap();
 }